AccountingPackage/requirements.txt
emoji-2.4.0
